Added DB checking and resolved error on Linux
For some reason the error checking gave an error on my chromebook running XFCE Linux.
I changed it for now and will test more when I get home and have access to my
Windows machine and virtualboxes.

I also added some checking on models to see if the database is actually enabled
before we allowing querying and fetching